1，不修改三方代码；
git-svn-id: https://172.200.1.200/svn/ZH6XXX/trunk/rpcservice@769 f6c6f0ce-9066-2545-9745-105408040c16

+++ include/rapidjson/msjson/msjsonserialize.hpp
@@ -99,7 +99,7 @@ namespace MSRPC
 		template <>
 		void in_serialize(const float& tValue, rapidjson::Value& vCurNode)
 		{
-			vCurNode.SetFloat(tValue);
+			vCurNode.SetDouble(tValue);
 		}
 
 
@@ -220,7 +220,7 @@ namespace MSRPC
 		template <>
 		void in_serialize(const rapidjson::Value& vCurNode, float& tValue)
 		{
-			tValue = vCurNode.GetFloat();
+			tValue = vCurNode.GetDouble();
 		}
 
 		template <>
